Refactoring codebase
