+++ src/components/ImgUpload/index.jsx
@@ -1,15 +1,17 @@
 import "./index.css";
 
 import { Avatar, LinearProgress, TextField } from "@mui/material";
+
 import { FaChevronCircleLeft, FaChevronCircleRight } from "react-icons/fa";
+import { LazyLoadImage } from "react-lazy-load-image-component";
 import React, { useRef, useState } from "react";
-import { auth, db, handleMultiUpload } from "../../lib/firebase";
+import { Link } from "react-router-dom";
 
 import AnimatedButton from "../../reusableComponents/AnimatedButton";
-import Camera from "../Camera";
-import { LazyLoadImage } from "react-lazy-load-image-component";
-import { Link } from "react-router-dom";
 import Popup from "../../reusableComponents/Popup";
+import Camera from "./Camera";
+
+import { auth, db, handleMultiUpload } from "../../lib/firebase";
 import firebase from "firebase/compat/app";
 import { useSnackbar } from "notistack";
