+++ src/components/Navbar/index.jsx
@@ -20,15 +20,7 @@ function Navbar({ onClick, open, setOpen, user, setUser, setLogout }) {
   const navigate = useNavigate();
   const rowMode = useContext(RowModeContext);
 
-  const buttonStyle = {
-    background: "linear-gradient(40deg, #e107c1, #59afc7)",
-    borderRadius: "20px",
-    margin: "10px",
-    ":hover": {
-      background: "linear-gradient(-40deg, #59afc7, #e107c1)",
-    },
-  };
-
+  
   useEffect(() => {
     const unsubscribe = auth.onAuthStateChanged((authUser) => {
       if (authUser) {
@@ -67,13 +59,7 @@ function Navbar({ onClick, open, setOpen, user, setUser, setLogout }) {
 
       {user && (
         <>
-          <div
-            style={{
-              display: "flex",
-              flexDirection: "row",
-              justifyContent: "center",
-            }}
-          >
+          <div className="container">
             <div className="rowConvert" onClick={onClick}>
               <AiOutlineInsertRowAbove style={{ margin: "auto" }} size={30} />
             </div>
@@ -82,31 +68,13 @@ function Navbar({ onClick, open, setOpen, user, setUser, setLogout }) {
                 onClick={() => setOpen((cur) => !cur)}
                 color="secondary"
                 variant="contained"
-                sx={{ ...buttonStyle, marginRight: "10px" }}
+                className="button-style"
               >
                 <FaUserCircle fontSize="large" />
                 {open && (
-                  <Box
-                    backgroundColor="#fff"
-                    color="black"
-                    padding="2px"
-                    width="80px"
-                    position="absolute"
-                    borderRadius="4px"
-                    marginTop={16}
-                    marginRight={3}
-                    sx={{
-                      width: "fit-content",
-                      mt: "10rem",
-                      mr: "4rem",
-                      vertical: "top",
-                      border: "1px solid black",
-                    }}
-                  >
+                  <Box className="nav-menu">
                     <Box
-                      display="flex"
-                      padding="0.5rem"
-                      sx={{ cursor: "pointer" }}
+                      className="nav-menu-item"
                       onClick={() =>
                         navigate("/dummygram/profile", {
                           state: {
@@ -123,9 +91,7 @@ function Navbar({ onClick, open, setOpen, user, setUser, setLogout }) {
                     </Box>
                     <Divider />
                     <Box
-                      display="flex"
-                      padding="0.5rem"
-                      sx={{ cursor: "pointer" }}
+                      className="nav-menu-item"
                       onClick={() => navigate("/dummygram/favourites")}
                     >
                       <Typography fontFamily="serif" fontSize="1rem">
@@ -134,9 +100,7 @@ function Navbar({ onClick, open, setOpen, user, setUser, setLogout }) {
                     </Box>
                     <Divider />
                     <Box
-                      display="flex"
-                      padding="0.5rem"
-                      sx={{ cursor: "pointer" }}
+                      className="nav-menu-item"
                       onClick={() => setLogout(true)}
                     >
                       <Typography fontFamily="Poppins" fontSize="0.9rem">
