+++ src/components/Post/index.css
@@ -15,6 +15,17 @@
   transform: scale(1.01);
 }
 
+.avatar{
+  border: 2px solid royalblue;
+  cursor: pointer;
+  transition: all 0.2s;
+}
+.avatar:hover{
+  box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 17px 0px;
+  border: 2px solid black;
+  scale: 1.1;
+}
+
 .postColumn {
   max-width: 400px;
   margin-top: 20px;
