+++ src/components/Post/index.jsx
@@ -2,10 +2,8 @@ import "react-lazy-load-image-component/src/effects/blur.css";
 import "./index.css";
 
 import {
-  Avatar,
   Box,
   Button,
-  ClickAwayListener,
   Dialog,
   DialogActions,
   DialogContent,
@@ -13,66 +11,58 @@ import {
   DialogTitle,
   Grid,
   IconButton,
-  Menu,
-  MenuItem,
+  // Menu,
+  // MenuItem,
   Paper,
   styled,
   useMediaQuery,
 } from "@mui/material";
-import { Link, useNavigate } from "react-router-dom";
-import { doc, updateDoc } from "firebase/firestore";
 import { useEffect, useState } from "react";
-
-import Caption from "../Caption";
-import CommentIcon from "@mui/icons-material/Comment";
-import DeleteTwoToneIcon from "@mui/icons-material/DeleteTwoTone";
-import DialogBox from "../../reusableComponents/DialogBox";
-import EmojiPicker from "emoji-picker-react";
+import EmojiPicker, { Emoji } from "emoji-picker-react";
 import { FaSave } from "react-icons/fa";
-import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
-import FavoriteOutlinedIcon from "@mui/icons-material/FavoriteOutlined";
-import ImageSlider from "../../reusableComponents/ImageSlider";
-import MoreHorizOutlinedIcon from "@mui/icons-material/MoreHorizOutlined";
-import ReadMore from "../ReadMore";
-import ReplyRoundedIcon from "@mui/icons-material/ReplyRounded";
-import Scroll from "../../reusableComponents/Scroll";
-import SentimentSatisfiedAltOutlinedIcon from "@mui/icons-material/SentimentSatisfiedAltOutlined";
-import TextField from "@mui/material/TextField";
+
+import { doc, updateDoc } from "firebase/firestore";
 import { db } from "../../lib/firebase";
 import firebase from "firebase/compat/app";
+import PostHeader from "./PostHeader";
+import EmojiBox from "./EmojiBox";
+import ImgBox from "./ImgBox";
+import CommentBox from "./CommentBox";
+
+import CommentIcon from "@mui/icons-material/Comment";
+import DeleteTwoToneIcon from "@mui/icons-material/DeleteTwoTone";
+import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
+import FavoriteOutlinedIcon from "@mui/icons-material/FavoriteOutlined";
+import ReplyRoundedIcon from "@mui/icons-material/ReplyRounded";
+import SentimentSatisfiedAltOutlinedIcon from "@mui/icons-material/SentimentSatisfiedAltOutlined";
 import { red } from "@mui/material/colors";
-import { saveAs } from "file-saver";
-import useCreatedAt from "../../hooks/useCreatedAt";
-import { useSnackbar } from "notistack";
 import { useTheme } from "@mui/material/styles";
 
-const ITEM_HEIGHT = 48;
+import { DialogBox, ImageSlider, Scroll } from "../../reusableComponents"
+import { Caption, ReadMore } from "../index"
+import useCreatedAt from "../../hooks/useCreatedAt";
+import { useSnackbar } from "notistack";
 
 function Post(prop) {
   const { postId, user, post, shareModal, setLink, setPostText, rowMode } =
     prop;
-  const { username, caption, imageUrl, avatar, likecount, timestamp } = post;
+  const { caption, imageUrl, likecount, timestamp } = post;
 
   const [comments, setComments] = useState([]);
   const [comment, setComment] = useState("");
+  const [postData, setPostData] = useState(post)
   const [likesNo, setLikesNo] = useState(likecount ? likecount.length : 0);
-  const [anchorEl, setAnchorEl] = useState(null);
-  const [editCaption, setEditCaption] = useState(caption);
   const [showEmojis, setShowEmojis] = useState(false);
   const [Open, setOpen] = useState(false);
-  const [openEditCaption, setOpenEditCaption] = useState(false);
   const [isCommentOpen, setisCommentOpen] = useState(false);
-  const [readMore, setReadMore] = useState(false);
   const [deleteCommentID, setDeleteCommentID] = useState("");
   const [openToDeleteComment, setOpenToDeleteComment] = useState(false);
 
   const time = useCreatedAt(timestamp);
   const theme = useTheme();
-  const navigate = useNavigate();
   const fullScreen = useMediaQuery(theme.breakpoints.down("md"));
   const { enqueueSnackbar } = useSnackbar();
 
-  const open = Boolean(anchorEl);
   const docRef = doc(db, "posts", postId);
   useEffect(() => {
     let unsubscribe;
@@ -172,13 +162,6 @@ function Post(prop) {
     }));
   }
 
-  // const computeGridSize = (imagesLength, imageIndex) => {
-  //   if (imageIndex === imagesLength - 1 && (imageIndex + 1) % 2 !== 0) {
-  //     return 12;
-  //   }
-  //   return 6;
-  // };
-
   const postHasImages = postImages.some((image) => Boolean(image.imageUrl));
 
   const tempLikeCount = likecount ? [...likecount] : [];
@@ -216,49 +199,6 @@ function Post(prop) {
     }
   }
 
-  async function deletePost() {
-    await db.collection("posts").doc(postId).delete();
-  }
-
-  const handleClickOpen = () => {
-    setOpen(true);
-    setAnchorEl(null);
-  };
-
-  const handleDownload = () => {
-    const urlimg = JSON.parse(imageUrl)[0].imageUrl;
-    saveAs(urlimg, "image");
-  };
-
-  const handleClickOpenCaption = async () => {
-    setOpenEditCaption(true);
-  };
-
-  const handleClickClosedCaption = () => {
-    setEditCaption(caption);
-    setOpenEditCaption(false);
-  };
-
-  const handleSubmitCaption = async () => {
-    const taskDocRef = doc(db, "posts", postId);
-    try {
-      await updateDoc(taskDocRef, {
-        caption: editCaption,
-      });
-    } catch (err) {
-      alert(err);
-    }
-    setOpenEditCaption(false);
-  };
-
-  // const handleCommentOpen = () => {
-  //   setisCommentOpen(true);
-  // };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
   const handleCommentClose = () => {
     setisCommentOpen(false);
   };
@@ -267,218 +207,30 @@ function Post(prop) {
     setOpenToDeleteComment(false);
   };
 
-  const handleReadPost = () => {
-    setReadMore(!readMore);
-  };
-
   return (
     <div
       className={`${rowMode ? "post" : "postColumn"}`}
       style={{ boxShadow: "0px 0px 5px 1px rgba(0, 0, 0, 0.4)" }}
     >
-      <div className="post__header">
-        <Avatar
-          className="post__avatar"
-          alt={username}
-          src={avatar}
-          sx={{
-            bgcolor: "royalblue",
-            border: "2px solid transparent",
-            display: "flex",
-            justifyContent: "center",
-            alignItems: "center",
-            cursor: "pointer",
-            "&:hover": {
-              boxShadow: "rgba(100, 100, 111, 0.2) 0px 7px 17px 0px",
-              border: "2px solid black",
-              scale: "1.1",
-            },
-          }}
-          onClick={() => {
-            navigate("/dummygram/profile", {
-              state: {
-                name: username,
-                avatar: avatar,
-              },
-            });
-          }}
-        />
-        <Link
-          to={`/dummygram/posts/${postId}`}
-          style={{ textDecoration: "none" }}
-        >
-          <h3 className="post__username">{username}</h3>
-          <p>{time}</p>
-        </Link>
-        <div className="social__icon__last">
-          <IconButton
-            aria-label="more"
-            id="long-button"
-            aria-controls={open ? "long-menu" : undefined}
-            aria-expanded={open ? "true" : undefined}
-            aria-haspopup="true"
-            onClick={(event) => setAnchorEl(event.currentTarget)}
-            sx={{
-              color: "var(--color)",
-            }}
-          >
-            <MoreHorizOutlinedIcon />
-          </IconButton>
-          <Menu
-            id="long-menu"
-            MenuListProps={{
-              "aria-labelledby": "long-button",
-            }}
-            anchorEl={anchorEl}
-            open={open}
-            onClose={() => setAnchorEl(null)}
-            PaperProps={{
-              style: {
-                maxHeight: ITEM_HEIGHT * 4.5,
-                width: "20ch",
-              },
-            }}
-          >
-            {user && username === user.displayName && (
-              <MenuItem onClick={handleClickOpen}> Delete </MenuItem>
-            )}
-            {user && username === user.displayName && (
-              <MenuItem onClick={handleClickOpenCaption}> Edit </MenuItem>
-            )}
-            <MenuItem onClick={handleDownload}> Download </MenuItem>
-            <MenuItem
-              onClick={() => {
-                navigate("/dummygram/profile", {
-                  state: {
-                    name: username,
-                    avatar: avatar,
-                  },
-                });
-              }}
-            >
-              Visit Profile
-            </MenuItem>
-          </Menu>
-          <>
-            <Dialog
-              fullWidth
-              open={openEditCaption}
-              onClose={handleClickClosedCaption}
-            >
-              <DialogTitle>Change Caption</DialogTitle>
-              <DialogContent>
-                <TextField
-                  autoFocus
-                  margin="dense"
-                  id="name"
-                  label="Enter Your Caption"
-                  type="text"
-                  fullWidth
-                  variant="standard"
-                  onChange={(e) => setEditCaption(e.target.value)}
-                  value={editCaption}
-                />
-              </DialogContent>
-              <DialogActions>
-                <Button onClick={handleClickClosedCaption}>Cancel</Button>
-                <Button onClick={handleSubmitCaption}>Submit</Button>
-              </DialogActions>
-            </Dialog>
-          </>
-
-          <Dialog
-            fullScreen={fullScreen}
-            open={Open}
-            onClose={handleClose}
-            aria-labelledby="responsive-dialog-title"
-          >
-            <DialogTitle id="responsive-dialog-title">
-              {"Delete Post?"}
-            </DialogTitle>
-            <DialogContent>
-              <DialogContentText>
-                Are you sure you want to delete this post?
-              </DialogContentText>
-            </DialogContent>
-            <DialogActions>
-              <Button onClick={handleClose}>Cancel</Button>
-              <Button onClick={deletePost}>Delete</Button>
-            </DialogActions>
-          </Dialog>
-        </div>
-      </div>
+      <PostHeader 
+        user={user}
+        postData={postData}
+      />
       <div className="post__container">
-        {postHasImages ? (
-          <ImageSlider
-            slides={postImages}
-            isCommentBox={false}
-            doubleClickHandler={likesHandler}
-          />
-        ) : (
-          <div className="post__background">
-            {caption.length >= 700 && readMore === false ? (
-              <>
-                <p className="post_caption">
-                  <Caption caption={caption.substr(0, 700)} />
-                  <button
-                    className="post__btn"
-                    onClick={() => handleReadPost()}
-                  >
-                    ... Read More
-                  </button>
-                </p>
-              </>
-            ) : (
-              <>
-                <p className="post_caption">
-                  <Caption caption={caption} />
-                </p>
-                {caption.length >= 700 && (
-                  <button
-                    className="post__less_btn"
-                    onClick={() => handleReadPost()}
-                  >
-                    ... Read Less
-                  </button>
-                )}
-              </>
-            )}
-          </div>
-        )}
-        <div className="post__text">
-          {caption && postHasImages && (
-            <>
-              {/* <strong style={{ color: "royalblue" }}>{username} </strong> */}
-              <ReadMore caption={caption} />
-            </>
-          )}
-        </div>
+        <ImgBox 
+          postHasImages={postHasImages} 
+          postImages={postImages} 
+          likesHandler={likesHandler} 
+          caption={caption}
+        />
+        
         {user && (
           <form className="post__commentBox">
-            <ClickAwayListener onClickAway={() => setShowEmojis(false)}>
-              <div className="social__icon">
-                <div className="emoji__icon">
-                  <SentimentSatisfiedAltOutlinedIcon
-                    onClick={() => {
-                      setShowEmojis((val) => !val);
-                    }}
-                  />
-                </div>
-                {showEmojis && (
-                  <div id="picker">
-                    <EmojiPicker
-                      emojiStyle="native"
-                      height={330}
-                      searchDisabled
-                      onEmojiClick={onEmojiClick}
-                      previewConfig={{
-                        showPreview: false,
-                      }}
-                    />
-                  </div>
-                )}
-              </div>
-            </ClickAwayListener>
+            <EmojiBox 
+              onEmojiClick={onEmojiClick} 
+              showEmojis={showEmojis}
+              setShowEmojis={setShowEmojis}
+            />
             <input
               className="post__input"
               type="text"
@@ -686,6 +438,7 @@ function Post(prop) {
                   </Grid>
                 </Grid>
               </Box>
+              <CommentBox setShowEmojis={setShowEmojis} showEmojis={showEmojis}/>
 
               {user && (
                 <form className="post__commentBox">
