+++ src/components/index.js
@@ -1,0 +1,21 @@
+import ImgUpload from "./ImgUpload";
+import Navbar from "./Navbar";
+import Post from "./Post";
+import ReadMore from "./ReadMore";
+import SideBar from "./SideBar";
+import Caption from "./Caption";
+import Favorite from "./Favorite";
+import LandingAnimation from "./LandingAnimation";
+import NotFound from "./NotFound";
+
+export {
+    ImgUpload,
+    Navbar,
+    Post,
+    ReadMore,
+    SideBar,
+    Caption,
+    Favorite,
+    LandingAnimation,
+    NotFound,
+}
