+++ src/reusableComponents/AnimatedButton.jsx
@@ -1,5 +1,5 @@
 import { Button } from "@mui/material";
-import Loader from "../components/Loader";
+import Loader from "../reusableComponents/Loader";
 
 export default function AnimatedButton(props) {
   const { loading, children, ..._props } = props;
