+++ src/reusableComponents/index.js
@@ -1,0 +1,19 @@
+import DialogBox from "./DialogBox";
+import ImageSlider from "./ImageSlider";
+import Loader from "./Loader";
+import Popup from "./Popup";
+import AnimatedButton from "./AnimatedButton";
+import Scroll from "./Scroll";
+import ShareModal from "./ShareModal";
+import SnackBar from "./SnackBar";
+
+export {
+    DialogBox,
+    ImageSlider,
+    Loader,
+    Popup,
+    AnimatedButton,
+    Scroll,
+    ShareModal,
+    SnackBar,
+}
